+++ yash-semantics/CHANGELOG.md
@@ -13,7 +13,7 @@ Terminology: A _public dependency_ is one that’s exposed through this crate’
 public API (e.g., re-exported types).
 A _private dependency_ is used internally and not visible to downstream users.
 
-## [0.17.1] - Unreleased
+## [0.18.0] - Unreleased
 
 ### Changed
 
@@ -23,7 +23,7 @@ A _private dependency_ is used internally and not visible to downstream users.
   syntax take effect on subsequent input.
 - Public dependency versions:
     - yash-env 0.15.0 → 0.15.3
-    - yash-syntax 0.22.0 → 0.22.1
+    - yash-syntax 0.22.0 → 0.23.0
 
 ## [0.17.0] - 2026-06-11
 
@@ -557,7 +557,7 @@ A _private dependency_ is used internally and not visible to downstream users.
 
 - Initial implementation of the `yash-semantics` crate
 
-[0.17.1]: https://github.com/magicant/yash-rs/releases/tag/yash-semantics-0.17.1
+[0.18.0]: https://github.com/magicant/yash-rs/releases/tag/yash-semantics-0.18.0
 [0.17.0]: https://github.com/magicant/yash-rs/releases/tag/yash-semantics-0.17.0
 [0.16.0]: https://github.com/magicant/yash-rs/releases/tag/yash-semantics-0.16.0
 [0.15.0]: https://github.com/magicant/yash-rs/releases/tag/yash-semantics-0.15.0